Fix Eclipse warnings.
This change is only there to fix warning issued by Eclipse.
It should have absolutely no impact on the program logic.

+++ java/src/com/android/inputmethod/latin/Utils.java
@@ -648,6 +648,6 @@ public class Utils {
 
     /** Convert pixel to DIP */
     public static int dipToPixel(float scale, int dip) {
-        return (int) ((float) dip * scale + 0.5);
+        return (int) (dip * scale + 0.5);
     }
 }
